Add files via upload
1. **Starting Out:** So, I just started learning React, and I thought it would be cool to build something simple to practice. I heard about this thing called a Todo List, where you can keep track of tasks you need to do.

2. **Adding Tasks:** First, I learned how to make a form where you can type in a task and hit "Add" to put it on the list. It was kinda neat seeing the tasks show up one after another.

3. **Checking off Tasks:** Then, I figured out how to add checkboxes next to each task. When you click one, it puts a line through the task to show it's done. That was satisfying!

4. **Removing Tasks:** Next, I learned how to add a "Remove" button to each task. Clicking it makes the task disappear from the list. It's nice having the option to get rid of stuff you don't need anymore.

5. **Filtering Tasks:** Lastly, I learned about filtering. I added buttons to show all tasks, only active tasks (ones not done yet), or completed tasks. It's handy for organizing the list.

This Todo List app was my first project with React, and it was a great way to learn the basics. I learned how to create a form for adding tasks, use checkboxes for marking tasks as done, and add buttons for removing tasks and filtering the list. It's a simple app, but it taught me a lot about how React works and how to build interactive components. Excited to keep learning and building more stuff with React!